+++ http4k-contract/src/test/kotlin/org/http4k/contract/DslContractRoutingHttpHandlerTest.kt
@@ -1,9 +1,18 @@
 package org.http4k.contract
 
+import com.natpryce.hamkrest.assertion.assertThat
 import org.http4k.contract.simple.SimpleJson
+import org.http4k.core.Method.GET
+import org.http4k.core.Request
+import org.http4k.core.Response
+import org.http4k.core.Status.Companion.OK
 import org.http4k.core.then
 import org.http4k.filter.ServerFilters
 import org.http4k.format.Jackson
+import org.http4k.hamkrest.hasStatus
+import org.http4k.routing.routes
+import org.http4k.routing.bind
+import org.junit.jupiter.api.Test
 
 class FunctionContractRoutingHttpHandlerTest : ContractRoutingHttpHandlerContract() {
     @Suppress("DEPRECATION")
@@ -23,4 +32,16 @@ class DslContractRoutingHttpHandlerTest : ContractRoutingHttpHandlerContract() {
                     routes += contractRoutes.toList()
                 }
             )
+
+    @Test
+    fun `can bind under a route match`() {
+        val app =
+            routes("/hello" bind routes("/there" bind
+                contract {
+                    renderer = SimpleJson(Jackson)
+                    routes += validPath bindContract GET to { Response(OK) }
+                })
+            )
+        assertThat(app(Request(GET, "/hello/there$validPath")), hasStatus(OK))
+    }
 }