+++ http4k-core/src/main/kotlin/org/http4k/lens/extensions.kt
@@ -3,6 +3,8 @@ package org.http4k.lens
 import dev.forkhandles.result4k.Failure
 import dev.forkhandles.result4k.Result
 import dev.forkhandles.result4k.Success
+import org.http4k.core.Request
+import org.http4k.routing.asRouter
 
 /**
  * Convert the result of a lens extraction to a Result4k type which
@@ -14,3 +16,15 @@ fun <IN, OUT> LensExtractor<IN, OUT>.asResult(): LensExtractor<IN, Result<OUT, L
         Failure(e)
     }
 }
+
+/**
+ * Check the content of any lens on a request for routing purposes.
+ */
+fun <T> Lens<Request, T>.matches(fn: (T) -> Boolean) = { r: Request ->
+    try {
+        fn(this(r))
+    } catch (e: LensFailure) {
+        false
+    }
+}.asRouter()
+