+++ http4k-opentelemetry/src/main/kotlin/org/http4k/filter/tracingFiltersOpenTelemetryExtensions.kt
@@ -73,7 +73,7 @@ fun ServerFilters.OpenTelemetryTracing(tracer: Tracer = Http4kOpenTelemetry.trac
                     try {
                         val ref = AtomicReference(next(req))
                         textMapPropagator.inject(Context.current(), ref, setter)
-                        ref.get()
+                        ref.get().also { setAttribute("http.status_code", it.status.code.toString()) }
                     } catch (t: Throwable) {
                         setStatus(ERROR, error(req, t))
                         throw t